requirements.txt
# Web Framework
streamlit

# Azure Services
azure-storage-blob
azure-search-documents
azure-identity

# AI/ML
openai

# Document Processing
python-docx
PyPDF2
pdfplumber
openpyxl

# Data Processing
pandas
numpy

# Configuration
python-dotenv

# HTTP Requests
requests

